+++ frontend/app/src/components/App.svelte
@@ -42,6 +42,7 @@
     import ActiveCall from "./home/video/ActiveCall.svelte";
     import { incomingVideoCall } from "../stores/video";
     import IncomingCall from "./home/video/IncomingCall.svelte";
+
     overrideItemIdKeyNameBeforeInitialisingDndZones("_id");
 
     const logger = inititaliseLogger(
@@ -124,6 +125,7 @@
             setDiamondMembershipFees,
             stakeNeuronForSubmittingProposals,
             updateMarketMakerConfig,
+            setPrincipalMigrationJobEnabled,
             pauseEventLoop: () => client.pauseEventLoop(),
             resumeEventLoop: () => client.resumeEventLoop(),
         };
@@ -326,6 +328,16 @@
                 console.log("Market maker config updated");
             } else {
                 console.log("Failed to update market maker config", resp);
+            }
+        });
+    }
+
+    function setPrincipalMigrationJobEnabled(enabled: boolean): void {
+        client.setPrincipalMigrationJobEnabled(enabled).then((resp) => {
+            if (resp) {
+                console.log("Principal migration job updated");
+            } else {
+                console.log("Failed to update principal migration job");
             }
         });
     }