+++ scripts/download-canister-wasm-dfx.sh
@@ -4,13 +4,15 @@ SCRIPT=$(readlink -f "$0")
 SCRIPT_DIR=$(dirname "$SCRIPT")
 cd $SCRIPT_DIR/..
 
-URL=$(jq '.canisters.event_store.wasm' dfx.json)
+CANISTER_NAME=$1
+
+URL=$(jq ".canisters.${CANISTER_NAME}.wasm" dfx.json)
 URL=$(echo "$URL" | tr -d '"')
 
 mkdir -p wasms
 cd wasms
 
-echo "Downloading event_store wasm"
+echo "Downloading $CANISTER_NAME wasm"
 
 HTTP_CODE=$(curl -sOL $URL --write-out "%{http_code}")