+++ backend/canisters/cycles_dispenser/CHANGELOG.md
@@ -5,6 +5,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [unreleased]
 
+### Added
+
+- Automatically add SNS controlled canisters to the whitelist ([#5625](https://github.com/open-chat-labs/open-chat/pull/5625))
+
+### Changed
+
+- Top up all canisters registered with the SNS ([#5621](https://github.com/open-chat-labs/open-chat/pull/5621))
+
 ## [[2.0.1092](https://github.com/open-chat-labs/open-chat/releases/tag/v2.0.1092-cycles_dispenser)] - 2024-03-07
 
 ### Changed